% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crd_data}
\alias{crd_data}
\title{Completely Randomized Design (CRD) Dataset}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{Variety}{Factor with 3 levels: Acorn, Butternut, Spaghetti}
\item{Replicate}{Factor indicating the replicate (rep1, rep2, rep3)}
\item{Yield}{Numeric yield measurement}
}
}
\usage{
crd_data
}
\description{
Yield data from a completely randomized experiment comparing three squash
varieties (Acorn, Butternut, Spaghetti) with three replicates each.
}
\examples{
data(crd_data)
str(crd_data)
}
\keyword{datasets}
