% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{factorial_data}
\alias{factorial_data}
\title{Factorial Design Dataset}
\format{
A data frame with 36 rows and 3 variables:
\describe{
\item{Temperature}{Factor with 3 levels: Tmp_A, Tmp_B, Tmp_C}
\item{Fertilizer}{Factor with 3 levels: Frt_A, Frt_B, Frt_C}
\item{Yield}{Numeric yield measurement}
}
}
\usage{
factorial_data
}
\description{
Yield data from a two-factor experiment crossing three temperature levels
(Tmp_A, Tmp_B, Tmp_C) with three fertilizer types (Frt_A, Frt_B, Frt_C),
with four replicates per combination.
}
\examples{
data(factorial_data)
str(factorial_data)
}
\keyword{datasets}
