% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcbd_data}
\alias{rcbd_data}
\title{Randomized Complete Block Design (RCBD) Dataset}
\format{
A data frame with 24 rows and 3 variables:
\describe{
\item{Block}{Factor with 4 levels indicating the block}
\item{Treatment}{Factor with 6 ordered levels: X25, X50, X75, X100, X125, X150}
\item{Yield}{Numeric yield measurement}
}
}
\usage{
rcbd_data
}
\description{
Yield data from a randomized complete block experiment testing six fertilizer
levels (X25, X50, X75, X100, X125, X150) across four blocks.
}
\examples{
data(rcbd_data)
str(rcbd_data)
}
\keyword{datasets}
