% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{split_plot_data}
\alias{split_plot_data}
\title{Split-Plot Design Dataset}
\format{
A data frame with 108 rows and 4 variables:
\describe{
\item{Replicate}{Factor with 6 levels (X1--X6)}
\item{WholePlot}{Factor with 6 levels representing the whole-plot treatment}
\item{SubPlot}{Factor with 3 levels (d1, d2, d3) representing the sub-plot treatment}
\item{Yield}{Numeric yield measurement}
}
}
\usage{
split_plot_data
}
\description{
Yield data from a split-plot experiment with six whole-plot treatments,
three sub-plot treatments, and six replicates.
}
\examples{
data(split_plot_data)
str(split_plot_data)
}
\keyword{datasets}
